requirements.txt
hubspot-api-client>=8.0.0
neo4j>=5.0.0
python-dotenv>=0.19.0
requests>=2.31.0
tqdm>=4.65.0
pandas>=2.0.0
tenacity>=8.0.0
pytest>=8.2.0
